Ab patch branch (#117)
* upgrade tornado to 6.1 due to failure I found in destroy teams build.

* fix formatting

* s3 browser does not work yet with jupyterhub 1.3 , trying 1.2

* without s3 browser.

* Revert "without s3 browser."

This reverts commit 86664967

* back to 1.1.0 as this requires more significant work with jupyterhub image.

* fix jupyterlab version to 2.2.9

+++ cli/datamaker_cli/commands/destroy.py
@@ -85,7 +85,7 @@ def destroy(filename: str, teams_only: bool, keep_demo: bool, debug: bool) -> No
         ctx.progress(95)
 
         try:
-            if not teams_only:
+            if not teams_only and not keep_demo:
                 destroy_toolkit(manifest=manifest)
         except botocore.exceptions.ClientError as ex:
             error = ex.response["Error"]
